+++ HelloSwift/HelloSwift/HelloSwift-Bridging-Header.h
@@ -1,0 +1,1 @@
+#import "AppboyKit.h"